# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HalalPilot Orchestration API
  version: 0.1.0
  description: |
    Kontrak antara skill OpenClaw `halalpilot` (pemanggil) dan layanan Express+SQLite (pelaksana).
    Prinsip: agent memutuskan bahasa & alur percakapan; API memutuskan hukum (rules/*.yaml), status, dan jadwal.
    Semua endpoint di 127.0.0.1:3000, header `Authorization: Bearer ${HALALPILOT_API_TOKEN}`.
    Endpoint yang mengubah state menerima `Idempotency-Key` (disarankan skill mengisi `<aksi>:<umk>:<hash payload>`).
servers:
  - url: http://127.0.0.1:3000/api/v1
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
...
paths:
...
  /umk/{id}/chase/request:
    post:
      summary: Mulai pengejaran untuk dokumen yang kurang (membuat chase_task tahap 1–4 sesuai chase-policy.yaml)
      parameters: [{ name: id, in: path, required: true, schema: { type: string }, description: "id numerik UMK atau kode UMK-xxx" }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { dokumen: { type: array, items: { type: string } } } } } } }
      responses: { '201': { description: "{tasks_created}" } }
  /chase/{task_id}/sent:
    post:
      summary: Skill menandai pengingat sudah terkirim (dipanggil agent setelah mengirim pesan)
      parameters: [{ name: task_id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { description: OK } }
...
  /umk/{id}/dossier:
    post:
      summary: Susun dossier PDF + Manual SJPH dari data terkini; status → menunggu_review; memberi tahu pendamping via /hooks/agent
      parameters: [{ name: id, in: path, required: true, schema: { type: string }, description: "id numerik UMK atau kode UMK-xxx" }]
      responses: { '201': { description: "{dossier_id, versi, pdf_path, sha256, url_dashboard}" } }
  /dossier/{did}/pdf:
    get:
      summary: Unduh/lihat PDF dossier (inline). Aktor UMK hanya untuk dossier miliknya.
      parameters: [{ name: did, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { description: application/pdf }, '404': { description: dossier/file tidak ada } }
  /dossier/{did}/review:
    post:
      summary: Pendamping menyetujui atau mengembalikan (dipanggil skill saat pendamping membalas "setuju UMK-017" / "kembalikan UMK-017 <alasan>")
      parameters: [{ name: did, in: path, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [aksi, pendamping_telegram_id]
              properties:
                aksi: { type: string, enum: [setuju, kembalikan] }
                catatan: { type: string }
                pendamping_telegram_id: { type: string }
      responses:
        '200': { description: "setuju → umk.status=siap_unggah; kembalikan → status=dikembalikan + chase ke UMK" }
        '403': { description: Bukan pendamping koperasi ini }
...